requirements.txt
tk
tkinterdnd2
git+https://github.com/openai/whisper.git
librosa
soundfile
noisereduce
python-docx
ffmpeg-python
